use colon instead of comma on a Map
In the section where AudioService is getting started (with `AudioService.start()`), the parameter named `params` expects a Map but there's a `Set` provided in the example. Maps are made up of key-value pairs and between the key and the value, a colon needs to be used. However on Sets, one needs to use comma to separate the elements of the Set.