Handle intent and provide result if the intent is consumed
When a new intent arrives, there can be several listeners for them. In case if the intent was handled by AudioService it should be consumed, and other listeners should not receive it. Otherwise, we need to return `false` in listener's method result. In this case the intent will go to the next listener.